Separated core functionality from service
