+++ app/src/main/kotlin/com/glodanif/bluetoothchat/data/service/BluetoothConnectionService.kt
@@ -1,80 +1,34 @@
 package com.glodanif.bluetoothchat.data.service
 
 import android.app.Service
-import android.bluetooth.BluetoothAdapter
 import android.bluetooth.BluetoothDevice
-import android.bluetooth.BluetoothServerSocket
-import android.bluetooth.BluetoothSocket
 import android.content.Context
 import android.content.Intent
 import android.content.ServiceConnection
-import android.graphics.BitmapFactory
 import android.os.Binder
 import android.os.Build
-import android.os.Environment
 import android.os.IBinder
 import android.support.v7.app.AppCompatActivity
-import com.glodanif.bluetoothchat.ChatApplication
 import com.glodanif.bluetoothchat.R
-import com.glodanif.bluetoothchat.data.database.ChatDatabase
 import com.glodanif.bluetoothchat.data.entity.ChatMessage
 import com.glodanif.bluetoothchat.data.entity.Conversation
-import com.glodanif.bluetoothchat.data.model.*
+import com.glodanif.bluetoothchat.data.model.OnConnectionListener
+import com.glodanif.bluetoothchat.data.model.OnFileListener
+import com.glodanif.bluetoothchat.data.model.OnMessageListener
 import com.glodanif.bluetoothchat.di.ComponentsManager
-import com.glodanif.bluetoothchat.ui.view.NotificationView
-import com.glodanif.bluetoothchat.ui.widget.ShortcutManager
-import com.glodanif.bluetoothchat.utils.Size
-import kotlinx.coroutines.experimental.CommonPool
-import kotlinx.coroutines.experimental.android.UI
-import kotlinx.coroutines.experimental.launch
 import java.io.File
-import java.io.IOException
-import java.util.*
 import javax.inject.Inject
 
-class BluetoothConnectionService : Service() {
+class BluetoothConnectionService : Service(), ConnectionSubject {
 
     private val binder = ConnectionBinder()
-    private val uiContext = UI
-    private val bgContext = CommonPool
-    private val adapter: BluetoothAdapter? = BluetoothAdapter.getDefaultAdapter()
-
-    private val APP_NAME = "BluetoothChat"
-    private val APP_UUID = UUID.fromString("220da3b2-41f5-11e7-a919-92ebcb67fe33")
-
-    enum class ConnectionState { CONNECTED, CONNECTING, NOT_CONNECTED, REJECTED, PENDING, LISTENING }
-    enum class ConnectionType { INCOMING, OUTCOMING }
 
     private var connectionListener: OnConnectionListener? = null
     private var messageListener: OnMessageListener? = null
     private var fileListener: OnFileListener? = null
 
-    private var acceptThread: AcceptJob? = null
-    private var connectThread: ConnectJob? = null
-    private var dataTransferThread: DataTransferThread? = null
-
-    @Volatile
-    private var connectionState: ConnectionState = ConnectionState.NOT_CONNECTED
-    @Volatile
-    private var connectionType: ConnectionType? = null
-
-    private var currentSocket: BluetoothSocket? = null
-    private var currentConversation: Conversation? = null
-    private var contract = Contract()
-
     @Inject
-    internal lateinit var db: ChatDatabase
-    @Inject
-    internal lateinit var preferences: UserPreferences
-    @Inject
-    internal lateinit var settings: SettingsManager
-
-    @Inject
-    internal lateinit var application: ChatApplication
-    @Inject
-    internal lateinit var shortcutManager: ShortcutManager
-    @Inject
-    internal lateinit var notificationView: NotificationView
+    internal lateinit var controller: ConnectionController
 
     override fun onBind(intent: Intent?): IBinder? {
         return binder
@@ -90,6 +44,7 @@ class BluetoothConnectionService : Service() {
     override fun onCreate() {
         super.onCreate()
         ComponentsManager.injectService(this)
+        controller.onNewForegroundMessage = { showNotification(it) }
         isRunning = true
     }
 
@@ -98,513 +53,156 @@ class BluetoothConnectionService : Service() {
         if (intent?.action == ACTION_STOP) {
 
             isRunning = false
-
-            connectionState = ConnectionState.NOT_CONNECTED
-            cancelConnections()
-            acceptThread?.cancel()
-            acceptThread = null
-
+            controller.stop()
             connectionListener?.onConnectionDestroyed()
 
             stopSelf()
             return START_NOT_STICKY
         }
 
-        prepareForAccept()
+        controller.prepareForAccept()
         showNotification(getString(R.string.notification__ready_to_connect))
         return Service.START_STICKY
     }
 
-    fun getCurrentConversation() = currentConversation
-
-    fun getCurrentContract() = contract
-
     private fun showNotification(message: String) {
-        val notification = notificationView.getForegroundNotification(message)
+        val notification = controller.createForegroundNotification(message)
         startForeground(FOREGROUND_SERVICE, notification)
     }
 
-    @Synchronized
-    fun disconnect() {
-        dataTransferThread?.cancel(true)
-        dataTransferThread = null
-        prepareForAccept()
-    }
-
-    @Synchronized
-    private fun prepareForAccept() {
-        cancelConnections()
-
-        if (isRunning) {
-            acceptThread = AcceptJob()
-            acceptThread?.start()
-            showNotification(getString(R.string.notification__ready_to_connect))
-        }
-    }
-
-    @Synchronized
     fun connect(device: BluetoothDevice) {
-
-        if (connectionState == ConnectionState.CONNECTING) {
-            connectThread?.cancel()
-            connectThread = null
-        }
-
-        dataTransferThread?.cancel(true)
-        acceptThread?.cancel()
-        dataTransferThread = null
-        acceptThread = null
-        currentSocket = null
-        currentConversation = null
-        contract.reset()
-        connectionType = null
-
-        connectThread = ConnectJob(device)
-        connectThread?.start()
-
-        launch(uiContext) { connectionListener?.onConnecting() }
+        controller.connect(device)
     }
 
-    @Synchronized
-    fun connected(socket: BluetoothSocket, type: ConnectionType) {
-
-        cancelConnections()
-
-        connectionType = type
-        currentSocket = socket
-
-        acceptThread?.cancel()
-        acceptThread = null
-
-        val transferEventsListener = object : DataTransferThread.TransferEventsListener {
-
-            override fun onMessageReceived(message: String) {
-                launch(uiContext) { this@BluetoothConnectionService.onMessageReceived(message) }
-            }
-
-            override fun onMessageSent(message: String) {
-                launch(uiContext) { this@BluetoothConnectionService.onMessageSent(message) }
-            }
-
-            override fun onConnectionPrepared(type: ConnectionType) {
-
-                showNotification(getString(R.string.notification__connected_to, socket.remoteDevice.name ?: "?"))
-                connectionState = ConnectionState.PENDING
-
-                if (type == ConnectionType.OUTCOMING) {
-                    contract.createConnectMessage(settings.getUserName(), settings.getUserColor()).let { message ->
-                        dataTransferThread?.write(message.getDecodedMessage())
-                    }
-                }
-            }
-
-            override fun onConnectionCanceled() {
-                currentSocket = null
-                currentConversation = null
-                contract.reset()
-            }
-
-            override fun onConnectionLost() {
-                connectionLost()
-            }
-        }
-
-        val fileEventsListener = object : DataTransferThread.OnFileListener {
-
-            override fun onFileSendingStarted(file: TransferringFile) {
-
-                fileListener?.onFileSendingStarted(file.name, file.size)
-
-                currentConversation?.let {
-
-                    val silently = application.currentChat != null && currentSocket != null &&
-                            application.currentChat.equals(currentSocket?.remoteDevice?.address)
-
-                    notificationView.showFileTransferNotification(it.displayName, it.deviceName,
-                            it.deviceAddress, file, 0, silently)
-                }
-            }
-
-            override fun onFileSendingProgress(file: TransferringFile, sentBytes: Long) {
-
-                launch(uiContext) {
-
-                    fileListener?.onFileSendingProgress(sentBytes, file.size)
-
-                    if (currentConversation != null) {
-                        notificationView.updateFileTransferNotification(sentBytes, file.size)
-                    }
-                }
-            }
-
-            override fun onFileSendingFinished(uid: Long, path: String) {
-
-                contract.createFileEndMessage().let { message ->
-                    dataTransferThread?.write(message.getDecodedMessage())
-                }
-
-                currentSocket?.let { socket ->
-
-                    val message = ChatMessage(socket.remoteDevice.address, Date(), true, "").apply {
-                        this.uid = uid
-                        seenHere = true
-                        messageType = PayloadType.IMAGE
-                        filePath = path
-                    }
-
-                    launch(bgContext) {
-
-                        val size = getImageSize(path)
-                        message.fileInfo = "${size.width}x${size.height}"
-                        message.fileExists = true
-
-                        db.messagesDao().insert(message)
-
-                        launch(uiContext) {
-
-                            fileListener?.onFileSendingFinished()
-                            messageListener?.onMessageSent(message)
-
-                            notificationView.dismissFileTransferNotification()
-                            currentConversation?.let {
-                                shortcutManager.addConversationShortcut(message.deviceAddress, it.displayName, it.color)
-                            }
-                        }
-                    }
-                }
-            }
-
-            override fun onFileSendingFailed() {
-
-                launch(uiContext) {
-                    fileListener?.onFileSendingFailed()
-                    notificationView.dismissFileTransferNotification()
-                }
-            }
-
-            override fun onFileReceivingStarted(file: TransferringFile) {
-
-                launch(uiContext) {
-
-                    fileListener?.onFileReceivingStarted(file.size)
-
-                    currentConversation?.let {
-
-                        val silently = application.currentChat != null && currentSocket != null &&
-                                application.currentChat.equals(currentSocket?.remoteDevice?.address)
-
-                        notificationView.showFileTransferNotification(it.displayName, it.deviceName,
-                                it.deviceAddress, file, 0, silently)
-                    }
-                }
-            }
-
-            override fun onFileReceivingProgress(file: TransferringFile, receivedBytes: Long) {
-
-                launch(uiContext) {
-
-                    fileListener?.onFileReceivingProgress(receivedBytes, file.size)
-
-                    if (currentConversation != null) {
-                        notificationView.updateFileTransferNotification(receivedBytes, file.size)
-                    }
-                }
-            }
-
-            override fun onFileReceivingFinished(uid: Long, path: String) {
-
-                currentSocket?.remoteDevice?.let { device ->
-
-                    val address = device.address
-                    val message = ChatMessage(address, Date(), false, "").apply {
-                        this.uid = uid
-                        messageType = PayloadType.IMAGE
-                        filePath = path
-                    }
-
-                    if (messageListener == null || application.currentChat == null || !application.currentChat.equals(address)) {
-                        notificationView.dismissMessageNotification()
-                        notificationView.showNewMessageNotification(getString(R.string.chat__image_message, "\uD83D\uDCCE"), currentConversation?.displayName,
-                                device.name, address, preferences.isSoundEnabled())
-                    } else {
-                        message.seenHere = true
-                    }
-
-                    launch(bgContext) {
-
-                        val size = getImageSize(path)
-                        message.fileInfo = "${size.width}x${size.height}"
-                        message.fileExists = true
-
-                        db.messagesDao().insert(message)
-
-                        launch(uiContext) {
-                            fileListener?.onFileReceivingFinished()
-                            messageListener?.onMessageReceived(message)
-
-                            notificationView.dismissFileTransferNotification()
-                            currentConversation?.let {
-                                shortcutManager.addConversationShortcut(address, it.displayName, it.color)
-                            }
-                        }
-                    }
-                }
-            }
-
-            override fun onFileReceivingFailed() {
-                launch(uiContext) {
-                    fileListener?.onFileReceivingFailed()
-                    notificationView.dismissFileTransferNotification()
-                }
-            }
-
-            override fun onFileTransferCanceled(byPartner: Boolean) {
-                launch(uiContext) {
-                    fileListener?.onFileTransferCanceled(byPartner)
-                    notificationView.dismissFileTransferNotification()
-                }
-            }
-        }
-
-        val eventsStrategy = TransferEventStrategy()
-        val filesDirectory = File(Environment.getExternalStorageDirectory(), getString(R.string.app_name))
-
-        dataTransferThread =
-                object : DataTransferThread(socket, type, transferEventsListener, filesDirectory, fileEventsListener, eventsStrategy) {
-                    override fun shouldRun(): Boolean {
-                        return isConnectedOrPending()
-                    }
-                }
-        dataTransferThread?.prepare()
-        dataTransferThread?.start()
-
-        launch(uiContext) { connectionListener?.onConnected(socket.remoteDevice) }
-    }
-
-    private fun getImageSize(path: String): Size {
-        val options = BitmapFactory.Options()
-        options.inJustDecodeBounds = true
-        BitmapFactory.decodeFile(path, options)
-        return Size(options.outWidth, options.outHeight)
-    }
-
-    @Synchronized
     fun stop() {
-        cancelConnections()
-        acceptThread?.cancel()
-        acceptThread = null
-        connectionState = ConnectionState.NOT_CONNECTED
+        controller.stop()
     }
 
-    private fun cancelConnections() {
-        connectThread?.cancel()
-        connectThread = null
-        dataTransferThread?.cancel()
-        dataTransferThread = null
-        currentSocket = null
-        currentConversation = null
-        contract.reset()
-        connectionType = null
+    fun disconnect() {
+        controller.disconnect()
     }
+
+    fun isConnected() = controller.isConnected()
+
+    fun isConnectedOrPending() = controller.isConnectedOrPending()
+
+    fun isPending() = controller.isPending()
+
+    fun getCurrentConversation(): Conversation? = controller.getCurrentConversation()
+
+    fun getTransferringFile(): TransferringFile? = controller.getTransferringFile()
+
+    fun getCurrentContract(): Contract = controller.getCurrentContract()
 
     fun sendMessage(message: Message) {
-
-        if (isConnectedOrPending()) {
-
-            val disconnect = message.type == Contract.MessageType.CONNECTION_REQUEST && !message.flag
-
-            dataTransferThread?.write(message.getDecodedMessage(), disconnect)
-
-            if (disconnect) {
-                dataTransferThread?.cancel(disconnect)
-                dataTransferThread = null
-                prepareForAccept()
-            }
-        }
-
-        if (message.type == Contract.MessageType.CONNECTION_RESPONSE) {
-            if (message.flag) {
-                connectionState = ConnectionState.CONNECTED
-            } else {
-                disconnect()
-            }
-            notificationView.dismissConnectionNotification()
-        }
+        controller.sendMessage(message)
     }
 
     fun sendFile(file: File, type: PayloadType) {
-
-        if (isConnected()) {
-            contract.createFileStartMessage(file, type).let { message ->
-                dataTransferThread?.write(message.getDecodedMessage())
-                dataTransferThread?.writeFile(message.uid, file)
-            }
-        }
-    }
-
-    fun getTransferringFile(): TransferringFile? {
-        return dataTransferThread?.getTransferringFile()
+        controller.sendFile(file, type)
     }
 
     fun cancelFileTransfer() {
-        dataTransferThread?.cancelFileTransfer()
-        notificationView.dismissFileTransferNotification()
+        controller.cancelFileTransfer()
     }
 
-    private fun onMessageSent(messageBody: String) = currentSocket?.let { socket ->
-
-        val message = Message(messageBody)
-        val sentMessage = ChatMessage(socket.remoteDevice.address, Date(), true, message.body)
-
-        if (message.type == Contract.MessageType.MESSAGE) {
-            sentMessage.seenHere = true
-            launch(bgContext) {
-                db.messagesDao().insert(sentMessage)
-                launch(uiContext) { messageListener?.onMessageSent(sentMessage) }
-                currentConversation?.let {
-                    shortcutManager.addConversationShortcut(sentMessage.deviceAddress, it.displayName, it.color)
-                }
-            }
-        }
-    }
-
-    private fun onMessageReceived(messageBody: String) {
-
-        val message = Message(messageBody)
-
-        if (message.type == Contract.MessageType.MESSAGE && currentSocket != null) {
-
-            handleReceivedMessage(message.uid, message.body)
-
-        } else if (message.type == Contract.MessageType.DELIVERY) {
-
-            if (message.flag) {
-                messageListener?.onMessageDelivered(message.uid)
-            } else {
-                messageListener?.onMessageNotDelivered(message.uid)
-            }
-        } else if (message.type == Contract.MessageType.SEEING) {
-
-            if (message.flag) {
-                messageListener?.onMessageSeen(message.uid)
-            }
-        } else if (message.type == Contract.MessageType.CONNECTION_RESPONSE) {
-
-            if (message.flag) {
-                handleConnectionApproval(message)
-            } else {
-                connectionState = ConnectionState.REJECTED
-                prepareForAccept()
-                connectionListener?.onConnectionRejected()
-            }
-        } else if (message.type == Contract.MessageType.CONNECTION_REQUEST && currentSocket != null) {
-
-            if (message.flag) {
-                handleConnectionRequest(message)
-            } else {
-                disconnect()
-                connectionListener?.onDisconnected()
-            }
-        } else if (message.type == Contract.MessageType.FILE_CANCELED) {
-            dataTransferThread?.cancelFileTransfer()
-        }
-    }
-
-    private fun handleReceivedMessage(uid: Long, text: String) = currentSocket?.let { socket ->
-
-        val device: BluetoothDevice = socket.remoteDevice
-
-        val receivedMessage = ChatMessage(device.address, Date(), false, text)
-        receivedMessage.uid = uid
-
-        if (messageListener == null || application.currentChat == null || !application.currentChat.equals(device.address)) {
-            notificationView.showNewMessageNotification(text, currentConversation?.displayName,
-                    device.name, device.address, preferences.isSoundEnabled())
-        } else {
-            receivedMessage.seenHere = true
-        }
-
-        launch(bgContext) {
-            db.messagesDao().insert(receivedMessage)
-            launch(uiContext) { messageListener?.onMessageReceived(receivedMessage) }
-            currentConversation?.let {
-                shortcutManager.addConversationShortcut(device.address, it.displayName, it.color)
-            }
-        }
-    }
-
-    private fun handleConnectionRequest(message: Message) = currentSocket?.let { socket ->
-
-        val device: BluetoothDevice = socket.remoteDevice
-
-        val parts = message.body.split("#")
-        val conversation = Conversation(device.address, device.name?: "?", parts[0], parts[1].toInt())
-
-        launch(bgContext) { db.conversationsDao().insert(conversation) }
-
-        currentConversation = conversation
-        contract setupWith if (parts.size >= 3) parts[2].trim().toInt() else 0
-
-        connectionListener?.onConnectedIn(conversation)
-
-        if (!application.isConversationsOpened && !(application.currentChat != null && application.currentChat.equals(device.address))) {
-            notificationView.showConnectionRequestNotification(
-                    "${conversation.displayName} (${conversation.deviceName})", preferences.isSoundEnabled())
-        }
-    }
-
-    private fun handleConnectionApproval(message: Message) = currentSocket?.let { socket ->
-
-        val device: BluetoothDevice = socket.remoteDevice
-
-        val parts = message.body.split("#")
-        val conversation = Conversation(device.address, device.name ?: "?", parts[0], parts[1].toInt())
-
-        launch(bgContext) { db.conversationsDao().insert(conversation) }
-
-        currentConversation = conversation
-        contract setupWith if (parts.size >= 3) parts[2].trim().toInt() else 0
-
-        connectionState = ConnectionState.CONNECTED
-        connectionListener?.onConnectionAccepted()
+    override fun handleConnectedOut(conversation: Conversation) {
         connectionListener?.onConnectedOut(conversation)
     }
 
-    private fun connectionFailed() {
-        currentSocket = null
-        currentConversation = null
-        contract.reset()
-        launch(uiContext) { connectionListener?.onConnectionFailed() }
-        connectionState = ConnectionState.NOT_CONNECTED
-        prepareForAccept()
+    override fun handleConnectedIn(conversation: Conversation) {
+        connectionListener?.onConnectedIn(conversation)
     }
 
-    private fun connectionLost() {
-
-        currentSocket = null
-        currentConversation = null
-        contract.reset()
-        if (isConnectedOrPending()) {
-            launch(uiContext) {
-                if (isPending() && connectionType == ConnectionType.INCOMING) {
-                    connectionState = ConnectionState.NOT_CONNECTED
-                    connectionListener?.onConnectionWithdrawn()
-                } else {
-                    connectionState = ConnectionState.NOT_CONNECTED
-                    connectionListener?.onConnectionLost()
-                }
-                prepareForAccept()
-            }
-        } else {
-            prepareForAccept()
-        }
+    override fun handleConnectionAccepted() {
+        connectionListener?.onConnectionAccepted()
     }
 
-    fun isConnected() = connectionState == ConnectionState.CONNECTED
+    override fun handleConnected(device: BluetoothDevice) {
+        connectionListener?.onConnected(device)
+    }
 
-    fun isPending() = connectionState == ConnectionState.PENDING
+    override fun handleConnectingInProgress() {
+        connectionListener?.onConnecting()
+    }
 
-    fun isConnectedOrPending() = isConnected() || isPending()
+    override fun handleDisconnected() {
+        connectionListener?.onDisconnected()
+    }
+
+    override fun handleConnectionRejected() {
+        connectionListener?.onConnectionRejected()
+    }
+
+    override fun handleConnectionFailed() {
+        connectionListener?.onConnectionFailed()
+    }
+
+    override fun handleConnectionLost() {
+        connectionListener?.onConnectionLost()
+    }
+
+    override fun handleConnectionWithdrawn() {
+        connectionListener?.onConnectionWithdrawn()
+    }
+
+    override fun handleFileSendingStarted(fileAddress: String?, fileSize: Long) {
+        fileListener?.onFileSendingStarted(fileAddress, fileSize)
+    }
+
+    override fun handleFileSendingProgress(sentBytes: Long, totalBytes: Long) {
+        fileListener?.onFileSendingProgress(sentBytes, totalBytes)
+    }
+
+    override fun handleFileSendingFinished() {
+        fileListener?.onFileSendingFinished()
+    }
+
+    override fun handleFileSendingFailed() {
+        fileListener?.onFileSendingFailed()
+    }
+
+    override fun handleFileReceivingStarted(fileSize: Long) {
+        fileListener?.onFileReceivingStarted(fileSize)
+    }
+
+    override fun handleFileReceivingProgress(sentBytes: Long, totalBytes: Long) {
+        fileListener?.onFileReceivingProgress(sentBytes, totalBytes)
+    }
+
+    override fun handleFileReceivingFinished() {
+        fileListener?.onFileReceivingFinished()
+    }
+
+    override fun handleFileReceivingFailed() {
+        fileListener?.onFileReceivingFailed()
+    }
+
+    override fun handleFileTransferCanceled(byPartner: Boolean) {
+        fileListener?.onFileTransferCanceled(byPartner)
+    }
+
+    override fun handleMessageReceived(message: ChatMessage) {
+        messageListener?.onMessageReceived(message)
+    }
+
+    override fun handleMessageSent(message: ChatMessage) {
+        messageListener?.onMessageSent(message)
+    }
+
+    override fun handleMessageDelivered(uid: Long) {
+        messageListener?.onMessageDelivered(uid)
+    }
+
+    override fun handleMessageNotDelivered(uid: Long) {
+        messageListener?.onMessageNotDelivered(uid)
+    }
+
+    override fun handleMessageSeen(uid: Long) {
+        messageListener?.onMessageSeen(uid)
+    }
+
+    override fun isAnybodyListeningForMessages() = messageListener != null
 
     fun setConnectionListener(listener: OnConnectionListener?) {
         this.connectionListener = listener
@@ -618,103 +216,10 @@ class BluetoothConnectionService : Service() {
         this.fileListener = listener
     }
 
-    private inner class AcceptJob {
-
-        private var serverSocket: BluetoothServerSocket? = null
-
-        init {
-            try {
-                serverSocket = adapter?.listenUsingRfcommWithServiceRecord(APP_NAME, APP_UUID)
-            } catch (e: IOException) {
-                e.printStackTrace()
-            }
-
-            connectionState = ConnectionState.LISTENING
-        }
-
-        fun start() = launch(bgContext) {
-
-            while (!isConnectedOrPending()) {
-
-                try {
-                    serverSocket?.accept()?.let { socket ->
-                        when (connectionState) {
-                            ConnectionState.LISTENING, ConnectionState.CONNECTING -> {
-                                connected(socket, ConnectionType.INCOMING)
-                            }
-                            ConnectionState.NOT_CONNECTED, ConnectionState.CONNECTED, ConnectionState.PENDING, ConnectionState.REJECTED -> try {
-                                socket.close()
-                            } catch (e: IOException) {
-                                e.printStackTrace()
-                            }
-                        }
-                    }
-                } catch (e: IOException) {
-                    e.printStackTrace()
-                    break
-                }
-            }
-        }
-
-        fun cancel() {
-            try {
-                serverSocket?.close()
-            } catch (e: IOException) {
-                e.printStackTrace()
-            }
-        }
-    }
-
-    private inner class ConnectJob(private val bluetoothDevice: BluetoothDevice) {
-
-        private var socket: BluetoothSocket? = null
-
-        fun start() = launch(bgContext) {
-
-            try {
-                socket = bluetoothDevice.createRfcommSocketToServiceRecord(APP_UUID)
-            } catch (e: IOException) {
-                e.printStackTrace()
-            }
-            connectionState = ConnectionState.CONNECTING
-
-            try {
-                socket?.connect()
-            } catch (connectException: IOException) {
-                connectException.printStackTrace()
-                try {
-                    socket?.close()
-                } catch (e: IOException) {
-                    e.printStackTrace()
-                }
-                connectionFailed()
-                return@launch
-            }
-
-            synchronized(this@BluetoothConnectionService) {
-                connectThread = null
-            }
-
-            socket?.let {
-                connected(it, ConnectionType.OUTCOMING)
-            }
-        }
-
-        fun cancel() {
-            try {
-                socket?.close()
-            } catch (e: IOException) {
-                e.printStackTrace()
-            }
-        }
-    }
-
     override fun onDestroy() {
         super.onDestroy()
         isRunning = false
-        cancelConnections()
-        acceptThread?.cancel()
-        acceptThread = null
+        controller.stop()
     }
 
     companion object {
