+++ app/src/main/kotlin/com/glodanif/bluetoothchat/data/service/ConnectionState.kt
@@ -1,0 +1,5 @@
+package com.glodanif.bluetoothchat.data.service
+
+enum class ConnectionState {
+    CONNECTED, CONNECTING, NOT_CONNECTED, REJECTED, PENDING, LISTENING
+}
