+++ app/src/main/kotlin/com/glodanif/bluetoothchat/data/service/ConnectionType.kt
@@ -1,0 +1,5 @@
+package com.glodanif.bluetoothchat.data.service
+
+enum class ConnectionType {
+    INCOMING, OUTCOMING
+}
