+++ app/src/main/kotlin/com/glodanif/bluetoothchat/data/service/DataTransferThread.kt
@@ -7,7 +7,7 @@ import kotlinx.coroutines.experimental.launch
 import java.io.*
 
 abstract class DataTransferThread(private val socket: BluetoothSocket,
-                                  private val type: BluetoothConnectionService.ConnectionType,
+                                  private val type: ConnectionType,
                                   private val transferListener: TransferEventsListener,
                                   private val filesDirectory: File,
                                   private val fileListener: OnFileListener,
@@ -344,7 +344,7 @@ abstract class DataTransferThread(private val socket: BluetoothSocket,
         fun onMessageReceived(message: String)
         fun onMessageSent(message: String)
 
-        fun onConnectionPrepared(type: BluetoothConnectionService.ConnectionType)
+        fun onConnectionPrepared(type: ConnectionType)
 
         fun onConnectionCanceled()
         fun onConnectionLost()
