+++ app/src/main/kotlin/com/glodanif/bluetoothchat/di/PerComponent.kt
@@ -4,4 +4,4 @@ import javax.inject.Scope
 
 @Scope
 @Retention(AnnotationRetention.RUNTIME)
-annotation class PerActivity
+annotation class PerComponent
