+++ app/src/main/kotlin/com/glodanif/bluetoothchat/di/component/ConversationsComponent.kt
@@ -1,11 +1,11 @@
 package com.glodanif.bluetoothchat.di.component
 
-import com.glodanif.bluetoothchat.di.PerActivity
+import com.glodanif.bluetoothchat.di.PerComponent
 import com.glodanif.bluetoothchat.di.module.ConversationsModule
 import com.glodanif.bluetoothchat.ui.activity.ConversationsActivity
 import dagger.Component
 
-@PerActivity
+@PerComponent
 @Component(dependencies = [(ApplicationComponent::class)], modules = [(ConversationsModule::class)])
 interface ConversationsComponent {
     fun inject(activity: ConversationsActivity)
