+++ app/src/main/kotlin/com/glodanif/bluetoothchat/di/component/ImagePreviewComponent.kt
@@ -1,11 +1,11 @@
 package com.glodanif.bluetoothchat.di.component
 
-import com.glodanif.bluetoothchat.di.PerActivity
+import com.glodanif.bluetoothchat.di.PerComponent
 import com.glodanif.bluetoothchat.di.module.*
 import com.glodanif.bluetoothchat.ui.activity.*
 import dagger.Component
 
-@PerActivity
+@PerComponent
 @Component(dependencies = [(ApplicationComponent::class)], modules = [(ImagePreviewModule::class)])
 interface ImagePreviewComponent {
     fun inject(activity: ImagePreviewActivity)
