+++ app/src/main/kotlin/com/glodanif/bluetoothchat/di/component/ReceivedImagesComponent.kt
@@ -1,11 +1,11 @@
 package com.glodanif.bluetoothchat.di.component
 
-import com.glodanif.bluetoothchat.di.PerActivity
+import com.glodanif.bluetoothchat.di.PerComponent
 import com.glodanif.bluetoothchat.di.module.ReceivedImagesModule
 import com.glodanif.bluetoothchat.ui.activity.ReceivedImagesActivity
 import dagger.Component
 
-@PerActivity
+@PerComponent
 @Component(dependencies = [(ApplicationComponent::class)], modules = [(ReceivedImagesModule::class)])
 interface ReceivedImagesComponent {
     fun inject(activity: ReceivedImagesActivity)
