+++ app/src/main/kotlin/com/glodanif/bluetoothchat/di/component/ScanComponent.kt
@@ -1,11 +1,11 @@
 package com.glodanif.bluetoothchat.di.component
 
-import com.glodanif.bluetoothchat.di.PerActivity
+import com.glodanif.bluetoothchat.di.PerComponent
 import com.glodanif.bluetoothchat.di.module.ScanModule
 import com.glodanif.bluetoothchat.ui.activity.ScanActivity
 import dagger.Component
 
-@PerActivity
+@PerComponent
 @Component(dependencies = [(ApplicationComponent::class)], modules = [(ScanModule::class)])
 interface ScanComponent {
     fun inject(activity: ScanActivity)
