+++ app/src/main/kotlin/com/glodanif/bluetoothchat/di/component/ServiceComponent.kt
@@ -1,12 +1,12 @@
 package com.glodanif.bluetoothchat.di.component
 
 import com.glodanif.bluetoothchat.data.service.BluetoothConnectionService
-import com.glodanif.bluetoothchat.di.module.ApplicationModule
+import com.glodanif.bluetoothchat.di.PerComponent
+import com.glodanif.bluetoothchat.di.module.ServiceModule
 import dagger.Component
-import javax.inject.Singleton
 
-@Singleton
-@Component(dependencies = [(ApplicationModule::class)])
+@PerComponent
+@Component(dependencies = [(ApplicationComponent::class)], modules = [(ServiceModule::class)])
 interface ServiceComponent {
     fun inject(service: BluetoothConnectionService)
 }
