+++ app/src/main/kotlin/com/glodanif/bluetoothchat/di/component/SettingsComponent.kt
@@ -1,13 +1,11 @@
 package com.glodanif.bluetoothchat.di.component
 
-import com.glodanif.bluetoothchat.di.PerActivity
-import com.glodanif.bluetoothchat.di.module.ProfileModule
+import com.glodanif.bluetoothchat.di.PerComponent
 import com.glodanif.bluetoothchat.di.module.SettingsModule
-import com.glodanif.bluetoothchat.ui.activity.ProfileActivity
 import com.glodanif.bluetoothchat.ui.activity.SettingsActivity
 import dagger.Component
 
-@PerActivity
+@PerComponent
 @Component(dependencies = [(ApplicationComponent::class)], modules = [(SettingsModule::class)])
 interface SettingsComponent {
     fun inject(activity: SettingsActivity)
