+++ app/src/main/kotlin/com/glodanif/bluetoothchat/di/module/ContactChooserModule.kt
@@ -1,7 +1,7 @@
 package com.glodanif.bluetoothchat.di.module
 
 import com.glodanif.bluetoothchat.data.model.ConversationsStorage
-import com.glodanif.bluetoothchat.di.PerActivity
+import com.glodanif.bluetoothchat.di.PerComponent
 import com.glodanif.bluetoothchat.ui.activity.ContactChooserActivity
 import com.glodanif.bluetoothchat.ui.presenter.ContactChooserPresenter
 import com.glodanif.bluetoothchat.ui.viewmodel.converter.ContactConverter
@@ -12,6 +12,6 @@ import dagger.Provides
 class ContactChooserModule(private val activity: ContactChooserActivity) {
 
     @Provides
-    @PerActivity
+    @PerComponent
     internal fun providePresenter(storage: ConversationsStorage, converter: ContactConverter): ContactChooserPresenter = ContactChooserPresenter(activity, storage, converter)
 }
