+++ app/src/main/kotlin/com/glodanif/bluetoothchat/di/module/ConversationsModule.kt
@@ -1,10 +1,9 @@
 package com.glodanif.bluetoothchat.di.module
 
 import com.glodanif.bluetoothchat.data.model.BluetoothConnector
-import com.glodanif.bluetoothchat.data.model.BluetoothConnectorImpl
 import com.glodanif.bluetoothchat.data.model.ConversationsStorage
 import com.glodanif.bluetoothchat.data.model.SettingsManager
-import com.glodanif.bluetoothchat.di.PerActivity
+import com.glodanif.bluetoothchat.di.PerComponent
 import com.glodanif.bluetoothchat.ui.activity.ConversationsActivity
 import com.glodanif.bluetoothchat.ui.presenter.ConversationsPresenter
 import com.glodanif.bluetoothchat.ui.viewmodel.converter.ConversationConverter
@@ -15,7 +14,7 @@ import dagger.Provides
 class ConversationsModule(private val activity: ConversationsActivity) {
 
     @Provides
-    @PerActivity
+    @PerComponent
     internal fun providePresenter(connector: BluetoothConnector, storage: ConversationsStorage,
                                   settings: SettingsManager, converter: ConversationConverter): ConversationsPresenter =
             ConversationsPresenter(activity, connector, storage, settings, converter)
