+++ app/src/main/kotlin/com/glodanif/bluetoothchat/di/module/ImagePreviewModule.kt
@@ -1,7 +1,7 @@
 package com.glodanif.bluetoothchat.di.module
 
 import com.glodanif.bluetoothchat.data.model.MessagesStorage
-import com.glodanif.bluetoothchat.di.PerActivity
+import com.glodanif.bluetoothchat.di.PerComponent
 import com.glodanif.bluetoothchat.ui.activity.ImagePreviewActivity
 import com.glodanif.bluetoothchat.ui.presenter.ImagePreviewPresenter
 import dagger.Module
@@ -12,6 +12,6 @@ import java.io.File
 class ImagePreviewModule(private val messageId: Long, private val image: File, private val activity: ImagePreviewActivity) {
 
     @Provides
-    @PerActivity
+    @PerComponent
     internal fun providePresenter(messages: MessagesStorage): ImagePreviewPresenter = ImagePreviewPresenter(messageId, image, activity, messages)
 }
