+++ app/src/main/kotlin/com/glodanif/bluetoothchat/di/module/ProfileModule.kt
@@ -3,7 +3,7 @@ package com.glodanif.bluetoothchat.di.module
 import com.glodanif.bluetoothchat.data.model.BluetoothScanner
 import com.glodanif.bluetoothchat.data.model.BluetoothScannerImpl
 import com.glodanif.bluetoothchat.data.model.SettingsManager
-import com.glodanif.bluetoothchat.di.PerActivity
+import com.glodanif.bluetoothchat.di.PerComponent
 import com.glodanif.bluetoothchat.ui.activity.ProfileActivity
 import com.glodanif.bluetoothchat.ui.presenter.ProfilePresenter
 import dagger.Module
@@ -13,10 +13,10 @@ import dagger.Provides
 class ProfileModule(private val activity: ProfileActivity) {
 
     @Provides
-    @PerActivity
+    @PerComponent
     internal fun providePresenter(settings: SettingsManager, scanner: BluetoothScanner): ProfilePresenter = ProfilePresenter(activity, settings, scanner)
 
     @Provides
-    @PerActivity
+    @PerComponent
     internal fun provideScanner(): BluetoothScanner = BluetoothScannerImpl(activity)
 }
