+++ app/src/main/kotlin/com/glodanif/bluetoothchat/di/module/ReceivedImagesModule.kt
@@ -1,7 +1,7 @@
 package com.glodanif.bluetoothchat.di.module
 
 import com.glodanif.bluetoothchat.data.model.MessagesStorage
-import com.glodanif.bluetoothchat.di.PerActivity
+import com.glodanif.bluetoothchat.di.PerComponent
 import com.glodanif.bluetoothchat.ui.activity.ReceivedImagesActivity
 import com.glodanif.bluetoothchat.ui.presenter.ReceivedImagesPresenter
 import dagger.Module
@@ -11,6 +11,6 @@ import dagger.Provides
 class ReceivedImagesModule(private val address: String?, private val activity: ReceivedImagesActivity) {
 
     @Provides
-    @PerActivity
+    @PerComponent
     internal fun providePresenter(messages: MessagesStorage): ReceivedImagesPresenter = ReceivedImagesPresenter(address, activity, messages)
 }
