+++ app/src/main/kotlin/com/glodanif/bluetoothchat/di/module/ScanModule.kt
@@ -1,7 +1,7 @@
 package com.glodanif.bluetoothchat.di.module
 
 import com.glodanif.bluetoothchat.data.model.*
-import com.glodanif.bluetoothchat.di.PerActivity
+import com.glodanif.bluetoothchat.di.PerComponent
 import com.glodanif.bluetoothchat.ui.activity.ScanActivity
 import com.glodanif.bluetoothchat.ui.presenter.ScanPresenter
 import dagger.Module
@@ -11,11 +11,11 @@ import dagger.Provides
 class ScanModule(private val activity: ScanActivity) {
 
     @Provides
-    @PerActivity
+    @PerComponent
     internal fun providePresenter(scanner: BluetoothScanner, connector: BluetoothConnector, fileManager: FileManager, preferences: UserPreferences): ScanPresenter =
             ScanPresenter(activity, scanner, connector, fileManager, preferences)
 
     @Provides
-    @PerActivity
+    @PerComponent
     internal fun provideScanner(): BluetoothScanner = BluetoothScannerImpl(activity)
 }
