+++ app/src/main/kotlin/com/glodanif/bluetoothchat/di/module/SettingsModule.kt
@@ -1,7 +1,7 @@
 package com.glodanif.bluetoothchat.di.module
 
 import com.glodanif.bluetoothchat.data.model.UserPreferences
-import com.glodanif.bluetoothchat.di.PerActivity
+import com.glodanif.bluetoothchat.di.PerComponent
 import com.glodanif.bluetoothchat.ui.activity.SettingsActivity
 import com.glodanif.bluetoothchat.ui.presenter.SettingsPresenter
 import dagger.Module
@@ -11,6 +11,6 @@ import dagger.Provides
 class SettingsModule(private val activity: SettingsActivity) {
 
     @Provides
-    @PerActivity
+    @PerComponent
     internal fun providePresenter(settings: UserPreferences): SettingsPresenter = SettingsPresenter(activity, settings)
 }
